runner/python-slave-runner/requirements.txt
# python-runner/requirements.txt
Flask==3.0.3
gitpython
virtualenv